fix: due to client/server runtime boundaries, moved to static export for client env values
