+++ src/sanity/lib/live.ts
@@ -1,6 +1,6 @@
 import { defineLive } from 'next-sanity';
 import { client } from './client';
-import { serverEnv } from '@/env/server';
+import { serverEnv } from '@/env/serverEnv';
 
 /**
  * Use defineLive to enable automatic revalidation and refreshing of your fetched content
