+++ src/utils/createEnv.ts
@@ -1,6 +1,10 @@
 import * as v from 'valibot';
 
 function createEnv<T extends v.ObjectEntries>(envSchema: T) {
+  if (typeof process === 'undefined') {
+    throw new Error('process is not available. This function should run in a Node.js environment.');
+  }
+
   type EnvKeys = keyof typeof envSchema;
   const envObj: Partial<Record<EnvKeys, string>> = Object.keys(envSchema).reduce(
     (acc, key) => {
